% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afplotadmin.r
\name{afplotadmin}
\alias{afplotadmin}
\title{plot admin levels}
\usage{
afplotadmin(country, datasource = "gadm", addcontext = FALSE)
}
\arguments{
\item{country}{a character vector of country names or iso3c character codes.}

\item{datasource}{data source, initial default 'gadm'}

\item{addcontext}{whether to plot neighbouring countries as well}
}
\value{
\code{sf}
}
\description{
*in progress
aim to plot admin levels for a single country
either on same map or facetted
to make it easier to see the structure
}
