% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afriadmin.r
\name{afriadmin}
\alias{afriadmin}
\title{Get admin polygons}
\usage{
afriadmin(country, level = "max", datasource = "gadm",
  plot = "mapview")
}
\arguments{
\item{country}{a character vector of country names or iso3c character codes.}

\item{level}{which admin level to return}

\item{datasource}{data source, initial default 'gadm'}

\item{plot}{option to display map 'mapview' for interactive, 'sf' for static}
}
\value{
\code{sf}
}
\description{
returns admin polygons for specified countries and optionally plots map
}
\examples{

sfnga2 <- afriadmin("nigeria", level=2)

}
