% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alllevels.r
\name{alllevels}
\alias{alllevels}
\title{get all admin levels into a single object}
\usage{
alllevels(country, datasource = "gadm", plot = "tmap")
}
\arguments{
\item{country}{a character vector of country names or iso3c character codes.}

\item{datasource}{data source, initial default 'gadm'}

\item{plot}{option to display map 'mapview' for interactive, 'sf' for static}
}
\value{
\code{sf}
}
\description{
*in progress
}
