% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareadmin.r
\name{compareadmin}
\alias{compareadmin}
\title{compare 2 sets of admin polygons for a country}
\usage{
compareadmin(
  country,
  level = c(1, 1),
  datasource = c("geoboundaries", "gadm"),
  type = c("simple", "simple"),
  version = c(NULL, NULL),
  quiet = FALSE,
  colors = c("red", "blue"),
  lwds = c(2, 2),
  col.regions = list(RColorBrewer::brewer.pal(9, "YlGn"), RColorBrewer::brewer.pal(9,
    "BuPu")),
  alpha = c(0.9, 0.9),
  layer.names = NULL,
  plotlegend = TRUE,
  alpha.regions = c(0.1, 0.1),
  plot = "mapview",
  plotshow = TRUE
)
}
\arguments{
\item{country}{a character vector of country names or iso3c character codes.}

\item{level}{1}

\item{datasource}{c('geoboundaries','gadm')}

\item{type}{for geoboundaries, boundary type defaults to 'simple' One of 'HPSCU', 'HPSCGS', 'SSCGS', 'SSCU', or 'precise' 'simple' 'precise standard' 'simple standard'
Determines the type of boundary link you receive. More on details}

\item{version}{for geoboundaries defaults to the most recent version of geoBoundaries available.
The geoboundaries version requested, with underscores. For example, 3_0_0 would return data
from version 3.0.0 of geoBoundaries.}

\item{quiet}{for geoboundaries, if TRUE no message while downloading and reading the data. Default to FALSE}

\item{plot}{option to display map 'mapview' for interactive, 'namestable' for a table of names, 'sf' for static}
}
\value{
\code{sf}
}
\description{
returns a mapview (or potentially later sf plot object), most arguments take a vector allowing two layers to be specified
}
\examples{

compareadmin("togo", level=2)
compareadmin("togo", level=2, datasource=c('geoboundaries','geoboundaries'), type=c('sscu','hpscu') )
#checking out problems with char encoding from rgeoboundaries
#e.g. togo admin2 accents malformed for sscu not hpscu
sfprecise <- rgeoboundaries::geoboundaries("togo", adm_lvl="adm2", type="hpscu")
sfsimple <- rgeoboundaries::geoboundaries("togo", adm_lvl="adm2", type="sscu")
#can correct simple version by specifying UTF8 within rgeob code
#sfsutf8 <- sf::st_read(path, quiet = quiet, options = "ENCODING=UTF8")
#but will that break the precise version ?
#yes it did break the precise version
#I should submit issue to the geoboundaries people


}
