% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{GADM36SF}
\alias{GADM36SF}
\title{admin levels available by country from GADM (and names of the levels)}
\format{
\code{dataframe}
}
\source{
\url{https://cran.r-project.org/web/packages/GADMTools/index.html}
}
\usage{
GADM36SF
}
\description{
admin levels available by country from GADM (and names of the levels)
}
\keyword{datasets}
