% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdxadmin.r
\name{hdxadmin}
\alias{hdxadmin}
\title{hdx admin layers starting to looking at downloading}
\usage{
hdxadmin(country)
}
\arguments{
\item{country}{a character vector of country names}
}
\value{
not sure yet
}
\description{
#todo vectorise
}
\examples{

hdxadmin("nigeria")

}
