% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdxlive.r
\name{hdxlive}
\alias{hdxlive}
\title{first go at interacting with hdx itos live admin layers
from https://github.com/SimonbJohnson/cod_topo}
\usage{
hdxlive(
  country,
  level = 2,
  colr_hdx = rgb(0, 0, 1, alpha = 0.4),
  colr_gadm = rgb(1, 0, 0, alpha = 0.4),
  plot = "compare"
)
}
\arguments{
\item{country}{a character vector of country names}

\item{level}{which admin level to return}

\item{colr_hdx}{colour for hdx boundaries in plots defaults blue semi-transparent}

\item{colr_gadm}{colour for gadm boundaries in plots defaults red semi-transparent}

\item{plot}{option to display map 'mapview' for interactive, 'sf' for static, 'compare' to sf compare with gadm}
}
\value{
not sure yet
}
\description{
NOTE I changed ZBM to ZMB for Zambia
}
\details{
start by checking whether I can open a few geojson files copied across
}
\examples{

hdxlive("angola")
hdxlive('benin',level=2)
hdxlive('zimbabwe',level=2)

}
