% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxadmin.r
\name{maxadmin}
\alias{maxadmin}
\title{return max admin level for a country}
\usage{
maxadmin(country, datasource = "gadm")
}
\arguments{
\item{country}{a character vector of country names}

\item{datasource}{data source, initial default 'gadm'}
}
\value{
integer vector of max admin levels
}
\description{
initialy gadm but could be other sources too
}
\details{
#todo vectorise
}
\examples{

maxlevel <- maxadmin("nigeria")

}
