% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afcountrynames.r
\name{afcountrynames}
\alias{afcountrynames}
\title{return all africa country names}
\usage{
afcountrynames(nameoriso3c = "name")
}
\arguments{
\item{nameoriso3c}{whether to return vector of 'name' or 'iso3c' or 'both'}
}
\value{
character vector of african country names
}
\description{
return all africa country names
}
\examples{

afcountrynames()

}
