% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{african_country_polygons}
\alias{african_country_polygons}
\alias{sf_af_gadm0}
\title{africa country polygons from GADM}
\format{
\code{sf}
}
\source{
\url{https://gadm.org/}
}
\usage{
sf_af_gadm0
}
\description{
africa country polygons from GADM
}
\keyword{datasets}
