% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdxadmin.r
\name{hdxadmin}
\alias{hdxadmin}
\title{hdx admin layers starting to looking at downloading
NOT WORKING YET}
\usage{
hdxadmin(country, level = 2, plot = "sf")
}
\arguments{
\item{country}{a character vector of country names}

\item{level}{which admin level to return}

\item{plot}{option to display map 'mapview' for interactive, 'sf' for static}
}
\value{
not sure yet
}
\description{
#todo vectorise
}
\examples{

#hdxadmin("nigeria")

}
